USB: get the rest of the files to compile.
The big goal here is that we'll have .h files in /sys/include and
library functions in /lib/libusb.a  so it's easy to build usb programs.

It all builds and as in all these conversion efforts I found errors
in the code.

The next step is probably to create /sys/include/usb/ and put
all .h files there, and make sure it all builds. Then we can
have /sys/src/cmds/usb/whatever and in there have example
usb programs.

The USB stuff was pretty incomplete so it's going to take a while
to get this rework right (assuming you all like it OK).

+++ sys/src/libusb/audio/audioctl.h
@@ -17,11 +17,11 @@ typedef struct Audiocontrol Audiocontrol;
 
 struct Audiocontrol {
 	char	*name;
-	uchar	readable;
-	uchar	settable;
-	uchar	chans;		/* 0 is master, non-zero is bitmap */
-	long	value[8];	/* 0 is master; value[0] == Undef -> all values Undef */
-	long	min, max, step;
+	uint8_t	readable;
+	uint8_t	settable;
+	uint8_t	chans;		/* 0 is master, non-zero is bitmap */
+	int32_t	value[8];	/* 0 is master; value[0] == Undef -> all values Undef */
+	int32_t	min, max, step;
 };
 
 extern Audiocontrol controls[2][Ncontrol];
@@ -32,7 +32,7 @@ extern int selectorid[2];
 extern int mixerid[2];
 extern int buttonendpt;
 
-int	ctlparse(char *s, Audiocontrol *c, long *v);
+int	ctlparse(char *s, Audiocontrol *c, int32_t *v);
 void	ctlevent(void);
 
 #pragma	varargck	type	"A"	Audiocontrol*
